fix(frontend): keep period parser out of Vue for backend typecheck
